+++ versions.tf
@@ -1,13 +1,15 @@
 terraform {
-  required_version = ">= 0.13"
+  required_version = ">= 1.1"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.96"
+      version = "~> 3.0"
     }
     azurecaf = {
       source  = "aztfmod/azurecaf"
       version = "~> 1.1"
     }
   }
+
+  experiments = [module_variable_optional_attrs]
 }